Database Refactor And Refactor of Entity Classes
