--- src/main/java/domain/salesman/dao/SalesmanDAO.java
@@ -1,5 +1,0 @@
-package domain.salesman.dao;
-
-public class SalesmanDAO {
-
-}
