+++ src/main/java/domain/user/dao/UserDAO.java
@@ -1,0 +1,4 @@
+package domain.user.dao;
+
+public class UserDAO {
+}
