+++ src/main/java/domain/user/salesman/SalesmanExistsException.java
@@ -1,4 +1,4 @@
-package domain.salesman;
+package domain.user.salesman;
 
 public class SalesmanExistsException extends Exception{
 }
