+++ src/main/java/domain/user/salesman/SalesmanNonMatchingPasswordException.java
@@ -1,4 +1,4 @@
-package domain.salesman;
+package domain.user.salesman;
 
 public class SalesmanNonMatchingPasswordException extends Exception{
 }
