+++ src/main/java/domain/user/salesman/SalesmanNotFoundException.java
@@ -1,4 +1,4 @@
-package domain.salesman;
+package domain.user.salesman;
 
 public class SalesmanNotFoundException extends Exception {
 }
