+++ src/main/java/domain/user/salesman/SalesmanPasswordVerifyException.java
@@ -1,4 +1,4 @@
-package domain.salesman;
+package domain.user.salesman;
 
 public class SalesmanPasswordVerifyException extends Exception{
 }
