+++ src/main/java/domain/user/salesman/dao/SalesmanDAO.java
@@ -1,0 +1,12 @@
+package domain.user.salesman.dao;
+
+import infrastructure.Database;
+
+public class SalesmanDAO {
+    private final Database database;
+
+
+    public SalesmanDAO(Database database) {
+        this.database = database;
+    }
+}
